// management/registry-ui/src/generated/client/models/model_Ticket.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type model_Ticket = {
    created_at?: string;
    customer_email?: string;
    customer_id?: number;
    id?: number;
    status?: string;
    subject?: string;
    updated_at?: string;
};